requirements.txt
lxml==4.7.1
matplotlib==3.5.1
numpy==1.22.2
Pillow==9.0.1
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm==4.62.3
pycocotools-windows==2.0.4
pycocotools==2.0.4
